# Disable gradle 7 module metadata publication